+++ kolibri/plugins/facility/assets/src/modules/userManagement/handlers.js
@@ -3,8 +3,10 @@ import { FacilityUserResource } from 'kolibri.resources';
 import samePageCheckGenerator from 'kolibri.utils.samePageCheckGenerator';
 import { _userState } from '../mappers';
 // An action for setting up the initial state of the app by fetching data from the server
-export function showUserPage(store, toRoute) {
-  store.dispatch('preparePage');
+export function showUserPage(store, toRoute, fromRoute) {
+  if (toRoute.name !== fromRoute.name) {
+    store.dispatch('preparePage');
+  }
   const facilityId = toRoute.params.facility_id || store.getters.activeFacilityId;
   const shouldResolve = samePageCheckGenerator(store);
   return FacilityUserResource.fetchCollection({