+++ kolibri/utils/main.py
@@ -223,18 +223,23 @@ def _upgrades_before_django_setup(updated, version):
 
 
 def _post_django_initialization():
-    if OPTIONS["Cache"]["CACHE_BACKEND"] != "redis":
-        try:
-            process_cache.cull()
-        except SQLite3DatabaseError:
-            shutil.rmtree(process_cache.directory, ignore_errors=True)
-            os.mkdir(process_cache.directory)
-            process_cache._cache = FanoutCache(
-                process_cache.directory,
-                settings.CACHES["process_cache"]["SHARDS"],
-                settings.CACHES["process_cache"]["TIMEOUT"],
-                **settings.CACHES["process_cache"]["OPTIONS"]
-            )
+    # Import here to prevent the module level access to Kolibri options
+    # which causes premature registration of Kolibri plugins.
+    from kolibri.deployment.default.cache import CACHES
+
+    if "process_cache" in CACHES:  # usually it means not using redis
+        if "DatabaseCache" not in CACHES["process_cache"]["BACKEND"]:
+            try:
+                process_cache.cull()
+            except SQLite3DatabaseError:
+                shutil.rmtree(process_cache.directory, ignore_errors=True)
+                os.mkdir(process_cache.directory)
+                process_cache._cache = FanoutCache(
+                    process_cache.directory,
+                    settings.CACHES["process_cache"]["SHARDS"],
+                    settings.CACHES["process_cache"]["TIMEOUT"],
+                    **settings.CACHES["process_cache"]["OPTIONS"]
+                )
 
 
 def _upgrades_after_django_setup(updated, version):